requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-jose
requests
paho-mqtt
psycopg2-binary
python-multipart
